pyproject.toml
[project]
name = "model-trainer"
version = "0.1.0"
description = "Browser-based image classifier trainer powered by Django + PyTorch"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0",
    "gunicorn>=25.1.0",
    "torch>=2.2",
    "torchvision>=0.17",
    "waitress>=3.0.2",
    "whitenoise>=6.11.0",
    "google-genai>=1.0.0",
    "PyPDF2>=3.0.0",
]

# Pull torch/torchvision from PyTorch's CPU-only index.
# The default PyPI torch is the CUDA build (~2.5 GB).
# The CPU build is ~700 MB — Railway has no GPU so CUDA is wasted space.
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
